+++ CHANGES.rst
@@ -4,6 +4,7 @@ USPTO PAIR Bulk Data API Client CHANGES
 
 development
 ===========
+- Implement the basic query and download parts of the USPTO PAIR Bulk Data API
 
 2017-10-19 0.0.0
 ================